+++ lib/recording_studio_pages/errors.rb
@@ -4,6 +4,7 @@ module RecordingStudioPages
   class Error < StandardError; end
   class DuplicateRegistration < Error; end
   class UnknownSectionType < Error; end
+  class UnknownCtaType < Error; end
   class UnknownTemplate < Error; end
   class InvalidSectionPayload < Error; end
   class UnpublishedPage < Error; end